+++ com/katas/domainobject/ToIdentifierType.java
@@ -1,4 +1,4 @@
-package com.katas;
+package com.katas.domainobject;
 
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
